requirements.txt
addict==2.4.0
basicsr==1.4.2
beautifulsoup4==4.12.2
caffe2==0.8.1
cityscapesscripts==2.2.2
diffusers==0.18.2
dominate==2.9.0
einops==0.7.0
external==0.0.1
fairscale==0.4.13
ftfy==6.1.3
fvcore==0.1.5.post20221221
hydra-core==1.3.2
imageio==2.33.0
imutils==0.5.4
iopath==0.1.10
kornia==0.7.0
lmdb==1.4.1
matplotlib==3.8.2
mc==0.0
mediapipe==0.10.8
mmdet==3.2.0
mmpose==1.2.0
numpy==1.24.3
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.16.3
opencv_python==4.8.1.78
openvino==2023.2.0
packaging==23.2
pandas==2.0.3
parrots==0.1.7
Pillow==9.3.0
Pillow==10.0.1
Pillow==10.1.0
prettytable==3.9.0
Pygments==2.17.2
pytorch_lightning==2.1.2
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.10.3
Requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
setuptools==68.0.0
Shapely==2.0.2
skimage==0.0
std_msgs==0.0.1
tabulate==0.9.0
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
termcolor==2.4.0
tifffile==2023.7.10
timm==0.6.7
torch==2.0.1+cu118
torch_tb_profiler==0.4.3
torchvision==0.15.2+cu118
tqdm==4.65.0
transformers==4.32.1
turbojpeg==0.0.2
yapf==0.40.2
